update ppl tests & IT tests
